[4.0.0] renamed argv module to cli
simple, communicates purpose more effectively and fewer conflicts
